Update inspectLink to support worker threads
- Necessary because of multiple await calls (and multiple calls to the api endpoint is possible)
